Time: 3 ms (34.48%) | Memory: 63 MB (48.93%) - LeetSync

+++ 11-container-with-most-water/container-with-most-water.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     int maxArea(vector<int>& height) {
         int n = height.size();
-        double maxArea = 0;
+        int maxArea = 0;
         int left = 0;
         int right = n - 1;
         while (left < right) {
-            double t = min(height[left], height[right]) * (right - left);
+            int t = min(height[left], height[right]) * (right - left);
             maxArea = max(maxArea, t);
             if (height[right] > height[left])
                 left++;
